BugFix,BalanceUpdate

+++ Assets/1.Main/2.Script/Controller/Player/PlayerSkillController.cs
@@ -27,7 +27,7 @@ public class PlayerSkillController : MonoBehaviour
     float DamageRigist;
     float HP;
     float KnockBackRate;
-    int Heal;
+    float Heal;
     int LastFire;
     int Pierce;
     int Boom;
