+++ Assets/1.Main/2.Script/Data/Status.cs
@@ -37,11 +37,11 @@ public struct Status
     public int Drain;
     public float Crush;
     public int Burn;
-    public int SkillHeal;
+    public float SkillHeal;
     public string KnickName;
     public string Title;
 
-    public Status(float hp,float hpMax,  float criRate, float criAttack,float atkSpeed, float damage, float defense, float speed, int maxammo, float reloadTime,float knockbackPer, float knockbackDist, float atkDist, int shotgun,int Level,float damagerigist,int lastfire, int pierce,int boom,float armorpierce,float remove,int drain,float crush,int burn,int heal,string knickName,string title)
+    public Status(float hp,float hpMax,  float criRate, float criAttack,float atkSpeed, float damage, float defense, float speed, int maxammo, float reloadTime,float knockbackPer, float knockbackDist, float atkDist, int shotgun,int Level,float damagerigist,int lastfire, int pierce,int boom,float armorpierce,float remove,int drain,float crush,int burn,float heal,string knickName,string title)
     {
         this.hp = hp;
         this.hpMax = hpMax; //이 객체의 최대체력과 현재 체력을 받아온 hp값으로 저장
